pyproject.toml
[project]
name = "llm-gateway"
version = "0.1.0"
description = "Minimal LLM gateway client for LiteLLM proxy"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "pydantic>=2.6",
  "openai>=1.30",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "sqlalchemy[asyncio]>=2.0",
  "aiosqlite>=0.19",
  "asyncpg>=0.29",
  "opentelemetry-api>=1.25",
  "opentelemetry-sdk>=1.25",
  "pytest>=8.1",
  "pytest-asyncio>=0.23",
  "mypy>=1.9",
  "ruff>=0.4",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B"]
